Cast long to double to prevent compiling problems

+++ MitoGraph.cxx
@@ -1308,9 +1308,12 @@ void GetTopologicalAttributes(vtkSmartPointer<vtkPolyData> PolyData, _mitoObject
         ne += (k==1) ? 1 : 0;
         nb += (k>=3) ? 1 : 0;
     }
-    attribute newAtt_1 = {"#End points",ne};
+
+    // @icaoberg casted long -> double
+    attribute newAtt_1 = {"#End points",(double)(ne)};
     mitoObject -> attributes.push_back(newAtt_1);
-    attribute newAtt_2 = {"#Bifurcations",nb};
+    // @icaoberg casted long -> double
+    attribute newAtt_2 = {"#Bifurcations",(double)(nb)};
     mitoObject -> attributes.push_back(newAtt_2);
     // NUMBER OF CONNECTED COMPONENTS
     vtkSmartPointer<vtkPolyDataConnectivityFilter> CC = vtkSmartPointer<vtkPolyDataConnectivityFilter>::New();
